chore: fix warnings

+++ packages/contracts/contracts/VolatilityFeed.sol
@@ -26,8 +26,8 @@ contract VolatilityFeed is AccessControl {
 	mapping(address => bool) public keeper;
 	// expiry array
 	uint256[] public expiries;
-	// interest rate
-	uint256 public interestRate;
+	// interest rate in e18 decimals
+	int256 public interestRate;
 
 	//////////////////////////
 	/// constant variables ///
@@ -119,9 +119,10 @@ contract VolatilityFeed is AccessControl {
 			_sabrParams.putVolvol
 		);
 	}
+
 	/// @notice set the interest rate that is used to compute forward price
 	function setInterestRate(
-		uint256 _interestRate
+		int256 _interestRate
 	) public {
 		_onlyGovernor();
 		interestRate = _interestRate;
@@ -157,11 +158,11 @@ contract VolatilityFeed is AccessControl {
 		if (sabrParams_.callAlpha == 0) {
 			revert CustomErrors.IVNotFound();
 		}
-		uint256 forwardPrice = underlyingPrice * (interestRate * time).exp2();
+		int256 forwardPrice = int256(underlyingPrice) * (interestRate * time).exp2();
 		if (!isPut) {
 			vol = SABR.lognormalVol(
 				int256(strikePrice),
-				int256(forwardPrice),
+				forwardPrice,
 				time,
 				sabrParams_.callAlpha * BIPS_SCALE,
 				sabrParams_.callBeta * BIPS_SCALE,
@@ -171,7 +172,7 @@ contract VolatilityFeed is AccessControl {
 		} else {
 			vol = SABR.lognormalVol(
 				int256(strikePrice),
-				int256(forwardPrice),
+				forwardPrice,
 				time,
 				sabrParams_.putAlpha * BIPS_SCALE,
 				sabrParams_.putBeta * BIPS_SCALE,
